requirements.txt
streamlit==1.32.0
numpy==1.26.0
pandas==2.0.0
matplotlib==3.7.1
altair==5.0.1
pillow==9.5.0
scikit-learn==1.3.2 